Decompose Test_Utility
