+++ tests/includes/test-class-static-press-ajax-init.php
@@ -10,25 +10,26 @@ namespace static_press\tests\includes;
 require_once dirname( __FILE__ ) . '/../testlibraries/class-array-url-handler.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-die-exception.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-expect-url.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-mock-creator.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url-creator.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url-handler.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-repository-for-test.php';
-require_once dirname( __FILE__ ) . '/../testlibraries/class-test-utility.php';
 use static_press\includes\Static_Press_Ajax_Init;
 use static_press\includes\Static_Press_Model_Url;
 use static_press\includes\Static_Press_Repository;
 use static_press\tests\testlibraries\Die_Exception;
+use static_press\tests\testlibraries\Mock_Creator;
 use static_press\tests\testlibraries\Model_Url;
+use static_press\tests\testlibraries\Model_Url_Creator;
 use static_press\tests\testlibraries\Model_Url_Handler;
 use static_press\tests\testlibraries\Repository_For_Test;
-use static_press\tests\testlibraries\Test_Utility;
 /**
  * StaticPress test case.
  *
  * @noinspection PhpUndefinedClassInspection
  */
 class Static_Press_Ajax_Init_Test extends \WP_UnitTestCase {
-	const DATE_FOR_TEST = '2019-12-23 12:34:56';
 	/**
 	 * Function get_urls() should trancate database table for list URL.
 	 * Function get_urls() should return urls of front page, static files, and SEO.
@@ -55,7 +56,7 @@ class Static_Press_Ajax_Init_Test extends \WP_UnitTestCase {
 			)
 		);
 		$expect_database = array(
-			Model_Url_Handler::create_model_url_fetched( 1, Static_Press_Model_Url::TYPE_OTHER_PAGE, '/test/', 1 ),
+			Model_Url_Creator::create_model_url_fetched( 1, Static_Press_Model_Url::TYPE_OTHER_PAGE, '/test/', 1 ),
 		);
 		$this->assertEquals( $expect_database, Repository_For_Test::get_all_url() );
 		ob_start();
@@ -80,8 +81,8 @@ class Static_Press_Ajax_Init_Test extends \WP_UnitTestCase {
 			null,
 			null,
 			new Static_Press_Repository(),
-			Test_Utility::create_remote_getter_mock(),
-			Test_Utility::create_terminator_mock()
+			Mock_Creator::create_remote_getter_mock(),
+			Mock_Creator::create_terminator_mock()
 		);
 		$reflection   = new \ReflectionClass( get_class( $static_press ) );
 		$method       = $reflection->getMethod( $method_name );
