+++ tests/includes/test-class-static-press-model-url-front-page.php
@@ -7,10 +7,10 @@
 
 namespace static_press\tests\includes;
 
-require_once dirname( __FILE__ ) . '/../testlibraries/class-test-utility.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-mock-creator.php';
 use static_press\includes\Static_Press_Model_Url;
 use static_press\includes\Static_Press_Model_Url_Front_Page;
-use static_press\tests\testlibraries\Test_Utility;
+use static_press\tests\testlibraries\Mock_Creator;
 /**
  * Reposistory test case.
  */
@@ -22,11 +22,11 @@ class Static_Press_Model_Url_Front_Page_Test extends \WP_UnitTestCase {
 		$expect         = array(
 			'type'          => Static_Press_Model_Url::TYPE_FRONT_PAGE,
 			'url'           => '/',
-			'last_modified' => Test_Utility::DATE_FOR_TEST,
+			'last_modified' => Mock_Creator::DATE_FOR_TEST,
 			'enable'        => null,
 		);
 		$url_front_page = new Static_Press_Model_Url_Front_Page(
-			Test_Utility::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' )
+			Mock_Creator::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' )
 		);
 		$this->assertEquals( $expect, $url_front_page->to_array() );
 	}
