+++ tests/includes/test-class-static-press-model-url-other.php
@@ -7,11 +7,10 @@
 
 namespace static_press\tests\includes;
 
-require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url.php';
-require_once dirname( __FILE__ ) . '/../testlibraries/class-test-utility.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-mock-creator.php';
 use static_press\includes\Static_Press_Model_Url;
 use static_press\includes\Static_Press_Model_Url_Other;
-use static_press\tests\testlibraries\Test_Utility;
+use static_press\tests\testlibraries\Mock_Creator;
 /**
  * StaticPress test case.
  */
@@ -20,12 +19,12 @@ class Static_Press_Model_Url_Other_Test extends \WP_UnitTestCase {
 	 * Constructor should set.
 	 */
 	public function test_constructor() {
-		$model_url_other = new Static_Press_Model_Url_Other( '/', Test_Utility::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' ) );
+		$model_url_other = new Static_Press_Model_Url_Other( '/', Mock_Creator::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' ) );
 		$this->assertEquals( '/', $model_url_other->get_url() );
 		$this->assertEquals(
 			array(
 				'type'          => Static_Press_Model_Url::TYPE_OTHER_PAGE,
-				'last_modified' => Test_Utility::DATE_FOR_TEST,
+				'last_modified' => Mock_Creator::DATE_FOR_TEST,
 				'url'           => '/',
 				'enable'        => null,
 			),
