+++ tests/includes/test-class-static-press-model-url.php
@@ -7,8 +7,9 @@
 
 namespace static_press\tests\includes;
 
+require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url-creator.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url-handler.php';
-require_once dirname( __FILE__ ) . '/../testlibraries/class-test-utility.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-mock-creator.php';
 
 use static_press\includes\Static_Press_Business_Logic_Exception;
 use static_press\includes\Static_Press_Model_Url;
@@ -16,8 +17,9 @@ use static_press\includes\Static_Press_Model_Url_Front_Page;
 use static_press\includes\Static_Press_Model_Url_Other;
 use static_press\includes\Static_Press_Model_Url_Seo;
 use static_press\includes\Static_Press_Model_Url_Static_File;
+use static_press\tests\testlibraries\Model_Url_Creator;
 use static_press\tests\testlibraries\Model_Url_Handler;
-use static_press\tests\testlibraries\Test_Utility;
+use static_press\tests\testlibraries\Mock_Creator;
 
 /**
  * StaticPress test case.
@@ -54,13 +56,13 @@ class Static_Press_Model_Url_Test extends \WP_UnitTestCase {
 	 * Function judge_to_dump() should be apporpriate string.
 	 */
 	public function provider_judge_to_dump() {
-		$date_time_factory_mock = Test_Utility::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' );
+		$date_time_factory_mock = Mock_Creator::create_date_time_factory_mock( 'create_date', 'Y-m-d h:i:s' );
 		return array(
 			array( new Static_Press_Model_Url_Seo( '', $date_time_factory_mock ), 1 ),
-			array( Model_Url_Handler::create_model_url_term(), 0 ),
+			array( Model_Url_Creator::create_model_url_term(), 0 ),
 			array( new Static_Press_Model_Url_Other( '', $date_time_factory_mock ), 1 ),
-			array( Model_Url_Handler::create_model_url_author(), 0 ),
-			array( Model_Url_Handler::create_model_url_single(), 0 ),
+			array( Model_Url_Creator::create_model_url_author(), 0 ),
+			array( Model_Url_Creator::create_model_url_single(), 0 ),
 			array( new Static_Press_Model_Url_Front_Page( $date_time_factory_mock ), 1 ),
 		);
 	}
