+++ tests/includes/test-class-static-press-static-file-judger.php
@@ -7,14 +7,18 @@
 
 namespace static_press\tests\includes;
 
-require_once dirname( __FILE__ ) . '/../testlibraries/class-test-utility.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-file-system-operator.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-mock-creator.php';
+require_once dirname( __FILE__ ) . '/../testlibraries/class-static-file-creator-for-test.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-theme-switcher.php';
 
 use Mockery;
 use static_press\includes\Static_Press_Model_Url;
 use static_press\includes\Static_Press_Model_Url_Static_File;
 use static_press\includes\Static_Press_Static_FIle_Judger;
-use static_press\tests\testlibraries\Test_Utility;
+use static_press\tests\testlibraries\File_System_Operator;
+use static_press\tests\testlibraries\Mock_Creator;
+use static_press\tests\testlibraries\Static_File_Creator_For_Test;
 use static_press\tests\testlibraries\Theme_Switcher;
 
 /**
@@ -44,12 +48,12 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 */
 	public function setUp() {
 		parent::setUp();
-		Test_Utility::create_static_file_readme();
-		Test_Utility::create_static_file_not_exist();
-		Test_Utility::create_static_file_not_updated();
-		Test_Utility::create_static_file_active_plugin();
-		Test_Utility::create_static_file_not_plugin_nor_theme();
-		Test_Utility::create_content_file_not_plugin_nor_theme();
+		Static_File_Creator_For_Test::create_static_file_readme();
+		Static_File_Creator_For_Test::create_static_file_not_exist();
+		Static_File_Creator_For_Test::create_static_file_not_updated();
+		Static_File_Creator_For_Test::create_static_file_active_plugin();
+		Static_File_Creator_For_Test::create_static_file_not_plugin_nor_theme();
+		Static_File_Creator_For_Test::create_content_file_not_plugin_nor_theme();
 		$this->theme_switcher->create_static_file_active_theme();
 		$this->theme_switcher->switch_theme();
 	}
@@ -58,9 +62,9 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * Puts up output directory, Mockery.
 	 */
 	public function tearDown() {
-		Test_Utility::delete_files();
-		Test_Utility::delete_files( ABSPATH . 'wp-content/uploads/' );
-		Test_Utility::delete_files( WP_CONTENT_DIR . '/app/uploads/' );
+		File_System_Operator::delete_files();
+		File_System_Operator::delete_files( ABSPATH . 'wp-content/uploads/' );
+		File_System_Operator::delete_files( WP_CONTENT_DIR . '/app/uploads/' );
 		Mockery::close();
 		parent::tearDown();
 	}
@@ -73,7 +77,7 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * @param integer                            $expect Expect.
 	 */
 	public function test_classify( $url, $expect ) {
-		$static_file_judger = new Static_Press_Static_FIle_Judger( Test_Utility::OUTPUT_DIRECTORY, Test_Utility::create_docuemnt_root_getter_mock() );
+		$static_file_judger = new Static_Press_Static_FIle_Judger( File_System_Operator::OUTPUT_DIRECTORY, Mock_Creator::create_docuemnt_root_getter_mock() );
 		$this->assertEquals( $expect, $static_file_judger->classify( $url ) );
 	}
 
@@ -86,9 +90,9 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 */
 	public function provider_classify() {
 		return array(
-			array( Test_Utility::create_static_file_readme(), 0 ),
-			array( Test_Utility::create_static_file_not_exist(), 0 ),
-			array( Test_Utility::create_static_file_active_plugin(), 1 ),
+			array( Static_File_Creator_For_Test::create_static_file_readme(), 0 ),
+			array( Static_File_Creator_For_Test::create_static_file_not_exist(), 0 ),
+			array( Static_File_Creator_For_Test::create_static_file_active_plugin(), 1 ),
 			array( $this->theme_switcher->create_static_file_active_theme(), 1 ),
 			array( $this->theme_switcher->create_static_file_theme_parent_activated(), 1 ),
 			array( $this->theme_switcher->create_static_file_non_active_theme(), 0 ),
@@ -99,7 +103,7 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * Function classify() should save as disable when dump directory is same with absolute path.
 	 */
 	public function test_classify_case_dump_directory_is_absolute_path() {
-		$static_file_judger = new Static_Press_Static_FIle_Judger( ABSPATH, Test_Utility::create_docuemnt_root_getter_mock() );
+		$static_file_judger = new Static_Press_Static_FIle_Judger( ABSPATH, Mock_Creator::create_docuemnt_root_getter_mock() );
 		$url                = new Static_Press_Model_Url_Static_File( Static_Press_Model_Url::TYPE_STATIC_FILE, '/', '/' );
 		$this->assertEquals( 0, $static_file_judger->classify( $url ) );
 	}
@@ -108,8 +112,8 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * Function classify() should save as disable when file is not updated after last dump.
 	 */
 	public function test_classify_case_non_update_file() {
-		$static_file_judger = new Static_Press_Static_FIle_Judger( Test_Utility::OUTPUT_DIRECTORY, Test_Utility::create_docuemnt_root_getter_mock() );
-		$url                = Test_Utility::create_static_file_not_updated();
+		$static_file_judger = new Static_Press_Static_FIle_Judger( File_System_Operator::OUTPUT_DIRECTORY, Mock_Creator::create_docuemnt_root_getter_mock() );
+		$url                = Static_File_Creator_For_Test::create_static_file_not_updated();
 		$this->assertEquals( 0, $static_file_judger->classify( $url ) );
 	}
 
@@ -117,8 +121,8 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * Function classify() should save as disable when URL is not activated plugin's static file.
 	 */
 	public function test_classify_case_non_active_plugin_static_file() {
-		$static_file_judger = new Static_Press_Static_FIle_Judger( Test_Utility::OUTPUT_DIRECTORY, Test_Utility::create_docuemnt_root_getter_mock() );
-		$url                = Test_Utility::create_static_file_non_active_plugin();
+		$static_file_judger = new Static_Press_Static_FIle_Judger( File_System_Operator::OUTPUT_DIRECTORY, Mock_Creator::create_docuemnt_root_getter_mock() );
+		$url                = Static_File_Creator_For_Test::create_static_file_non_active_plugin();
 		$this->assertEquals( 0, $static_file_judger->classify( $url ) );
 	}
 
@@ -129,7 +133,7 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 * @param Static_Press_Model_Url_Static_File $url URL.
 	 */
 	public function test_classify_case_static_file_not_plugin_nor_theme( $url ) {
-		$static_file_judger = new Static_Press_Static_FIle_Judger( Test_Utility::OUTPUT_DIRECTORY, Test_Utility::create_docuemnt_root_getter_mock() );
+		$static_file_judger = new Static_Press_Static_FIle_Judger( File_System_Operator::OUTPUT_DIRECTORY, Mock_Creator::create_docuemnt_root_getter_mock() );
 		$this->assertEquals( 1, $static_file_judger->classify( $url ) );
 	}
 
@@ -138,8 +142,8 @@ class Static_Press_Static_FIle_Jugder_Test extends \WP_UnitTestCase {
 	 */
 	public function provider_classify_case_static_file_not_plugin_nor_theme() {
 		return array(
-			array( Test_Utility::create_static_file_not_plugin_nor_theme() ),
-			array( Test_Utility::create_content_file_not_plugin_nor_theme() ),
+			array( Static_File_Creator_For_Test::create_static_file_not_plugin_nor_theme() ),
+			array( Static_File_Creator_For_Test::create_content_file_not_plugin_nor_theme() ),
 		);
 	}
 }
